Squashed 'airdcpp-webapi/' changes from c3f416c..0ab9a7b
0ab9a7b Revert "Update JSON" to maintain compatibility with GCC 4.8
7ce53d2 Fix a few issues reported by Valgrind
09b0671 Update JSON
f31764b Web server protocols can be disabled again
54c58f7 Remove support for sharing subdirectories of other shared directories in other share profiles
9932204 Update translation strings
a21704d Web server: add default ports, don't fire the "Started" event if the server failed to start, validate the (user) config before starting
62eb962 Include directory updates
b6a2a6a MSVC: use the included websocketpp

git-subtree-dir: airdcpp-webapi
git-subtree-split: 0ab9a7be84e759ab4fea175be2f4f181fed3038c

+++ web-server/FileServer.cpp
@@ -267,7 +267,7 @@ namespace webserver {
 		try {
 			File f(request, File::READ, File::OPEN);
 			f.setPos(startPos);
-			output_ = f.read(endPos + 1);
+			output_ = f.read(static_cast<size_t>(endPos) + 1);
 		} catch (const FileException& e) {
 			output_ = e.getError();
 			return websocketpp::http::status_code::not_found;
